+++ adcom1/display_format.go
@@ -29,7 +29,9 @@ type DisplayFormat struct {
 	// Definition:
 	//   Relative width of the creative when expressing size as a ratio.
 	//   Note that mixing absolute and relative sizes is not recommended.
-	WRatio uint64 `json:"wratio,omitempty"`
+	// Dev note:
+	//   This is kept as `uint8` because ratio values are expected to be quite small (like 16:9).
+	WRatio uint8 `json:"wratio,omitempty"`
 
 	// Attribute:
 	//   hratio
@@ -38,7 +40,9 @@ type DisplayFormat struct {
 	// Definition:
 	//   Relative height of the creative when expressing size as a ratio.
 	//   Note that mixing absolute and relative sizes is not recommended.
-	HRatio uint64 `json:"hratio,omitempty"`
+	// Dev note:
+	//   This is kept as `uint8` because ratio values are expected to be quite small (like 16:9).
+	HRatio uint8 `json:"hratio,omitempty"`
 
 	// Attribute:
 	//   expdir
